+++ imi/src/main/java/org/motechproject/nms/imi/web/RescheduleController.java
@@ -1,49 +1,35 @@
-package org.motechproject.nms.imi.service.impl;
+package org.motechproject.nms.imi.web;
 
 import org.motechproject.event.MotechEvent;
-import org.motechproject.event.listener.EventRelay;
 import org.motechproject.event.listener.annotations.MotechListener;
 import org.motechproject.nms.imi.domain.CallDetailRecord;
 import org.motechproject.nms.imi.domain.CallRetry;
 import org.motechproject.nms.imi.domain.CallStage;
 import org.motechproject.nms.imi.repository.CallRetryDataService;
 import org.motechproject.nms.imi.service.RequestId;
-import org.motechproject.nms.imi.service.ReschedulerService;
 import org.motechproject.nms.kilkari.domain.Subscription;
 import org.motechproject.nms.kilkari.service.SubscriptionService;
 import org.motechproject.nms.props.domain.DayOfTheWeek;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Controller;
 
-import java.util.HashMap;
-import java.util.Map;
+@Controller
+public class RescheduleController {
 
-@Service("reschedulerService")
-public class ReschedulerServiceImpl implements ReschedulerService {
+    private static final String RESCHEDULE_CALL = "nms.imi.reschedule_call";
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(ReschedulerServiceImpl.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(RescheduleController.class);
 
-    private EventRelay eventRelay;
     private SubscriptionService subscriptionService;
     private CallRetryDataService callRetryDataService;
 
 
     @Autowired
-    public ReschedulerServiceImpl(EventRelay eventRelay, SubscriptionService subscriptionService,
-                                  CallRetryDataService callRetryDataService) {
-        this.eventRelay = eventRelay;
+    public RescheduleController(SubscriptionService subscriptionService, CallRetryDataService callRetryDataService) {
         this.subscriptionService = subscriptionService;
         this.callRetryDataService = callRetryDataService;
-    }
-
-
-    public void sendRescheduleMessage(CallDetailRecord cdr) {
-        Map<String, Object> eventParams = new HashMap<>();
-        eventParams.put("CDR", cdr);
-        MotechEvent motechEvent = new MotechEvent(RESCHEDULE_CALL, eventParams);
-        eventRelay.sendEventMessage(motechEvent);
     }
