+++ imi/src/main/resources/META-INF/spring/blueprint.xml
@@ -15,9 +15,6 @@
 
     <osgi:service ref="settingsService" interface="org.motechproject.nms.imi.service.SettingsService" />
 
-    <osgi:service ref="reschedulerService" interface="org.motechproject.nms.imi.service.ReschedulerService" />
-
-
 
     <osgi:reference id="subscriptionService"
                     interface="org.motechproject.nms.kilkari.service.SubscriptionService" />
